Ongoing refactoring of sql client out of vector_server
