+++ python/hsfs/core/vector_server.py
@@ -13,27 +13,20 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 #
-import asyncio
 import io
 import logging
-import re
 
 import avro.io
 import avro.schema
 import numpy as np
 import pandas as pd
 import polars as pl
-from hsfs import client, feature_view, training_dataset, util
+from hsfs import client, feature_view, training_dataset
 from hsfs.core import (
-    feature_view_api,
     feature_view_engine,
     online_store_rest_client_engine,
-    storage_connector_api,
-    training_dataset_api,
     transformation_function_engine,
 )
-from hsfs.serving_key import ServingKey
-from sqlalchemy import bindparam, exc, sql, text
 
 
 _logger = logging.getLogger(__name__)
@@ -61,6 +54,7 @@ class VectorServer:
             features = []
         self._training_dataset_version = training_dataset_version
         self._features = [] if features is None else features
+
         self._feature_vector_col_name = (
             [
                 feat.name
@@ -75,24 +69,13 @@ class VectorServer:
             else []
         )
         self._skip_fg_ids = skip_fg_ids or set()
-        self._prepared_statement_engine = None
-        self._prepared_statements = None
-        self._helper_column_prepared_statements = None
+
         self._serving_keys = serving_keys
-        self._valid_serving_key = None
-        self._pkname_by_serving_index = None
-        self._prefix_by_serving_index = None
-        self._serving_key_by_serving_index = {}
-        self._external = True
         self._feature_store_id = feature_store_id
         self._feature_store_name = feature_store_name
         self._feature_view_name = feature_view_name
         self._feature_view_version = feature_view_version
-        self._training_dataset_api = training_dataset_api.TrainingDatasetApi(
-            feature_store_id
-        )
-        self._feature_view_api = feature_view_api.FeatureViewApi(feature_store_id)
-        self._storage_connector_api = storage_connector_api.StorageConnectorApi()
+
         self._transformation_function_engine = (
             transformation_function_engine.TransformationFunctionEngine(
                 feature_store_id
@@ -103,8 +86,8 @@ class VectorServer:
         )
         self._transformation_functions = None
         self._required_serving_keys = None
-        self._async_pool = None
 
+        self._online_store_sql_client = None
         self._online_store_rest_client_engine = None
         self._init_online_store_rest_client = None
         self._init_online_store_sql_client = None
@@ -164,7 +147,7 @@ class VectorServer:
 
         if self._init_online_store_sql_client is True:
             _logger.info("Initialising Vector Server Online SQL client")
-            self.init_prepared_statement(
+            self.online_store_sql_client.init_prepared_statement(
                 entity,
                 batch,
                 external,
@@ -178,157 +161,6 @@ class VectorServer:
     def init_transformation(self, entity):
         # attach transformation functions
         self._transformation_functions = self._get_transformation_fns(entity)
-
-    def init_prepared_statement(
-        self,
-        entity,
-        batch,
-        external,
-        inference_helper_columns,
-        options=None,
-    ):
-        if isinstance(entity, feature_view.FeatureView):
-            if inference_helper_columns:
-                helper_column_prepared_statements = (
-                    self._feature_view_api.get_serving_prepared_statement(
-                        entity.name, entity.version, batch, inference_helper_columns
-                    )
-                )
-            prepared_statements = self._feature_view_api.get_serving_prepared_statement(
-                entity.name, entity.version, batch, inference_helper_columns=False
-            )
-        elif isinstance(entity, training_dataset.TrainingDataset):
-            prepared_statements = (
-                self._training_dataset_api.get_serving_prepared_statement(entity, batch)
-            )
-        else:
-            raise ValueError(
-                "Object type needs to be `feature_view.FeatureView` or `training_dataset.TrainingDataset`."
-            )
-        # reset values to default, as user may be re-initialising with different parameters
-        self._prepared_statement_engine = None
-        self._prepared_statements = None
-        self._helper_column_prepared_statements = None
-        self._external = external
-        # create aiomysql engine
-        loop = asyncio.get_event_loop()
-        loop.run_until_complete(
-            self._set_aiomysql_connection(len(prepared_statements), options=options)
-        )
-
-        (
-            self._prepared_statements,
-            feature_name_order_by_psp,
-            prefix_by_serving_index,
-        ) = self._parametrize_prepared_statements(prepared_statements, batch)
-
-        if inference_helper_columns:
-            (
-                self._helper_column_prepared_statements,
-                _,
-                _,
-            ) = self._parametrize_prepared_statements(
-                helper_column_prepared_statements, batch
-            )
-
-        self._prefix_by_serving_index = prefix_by_serving_index
-        for sk in self._serving_keys:
-            self._serving_key_by_serving_index[sk.join_index] = (
-                self._serving_key_by_serving_index.get(sk.join_index, []) + [sk]
-            )
-        # sort the serving by PreparedStatementParameter.index
-        for join_index in self._serving_key_by_serving_index:
-            # feature_name_order_by_psp do not include the join index when the joint feature only contains label only
-            # But _serving_key_by_serving_index include the index when the join_index is 0 (left side)
-            if join_index in feature_name_order_by_psp:
-                self._serving_key_by_serving_index[join_index] = sorted(
-                    self._serving_key_by_serving_index[join_index],
-                    key=lambda _sk: feature_name_order_by_psp[join_index].get(
-                        _sk.feature_name, 0
-                    ),
-                )
-
-        self._valid_serving_key = set(
-            [sk.feature_name for sk in self._serving_keys]
-            + [sk.required_serving_key for sk in self._serving_keys]
-        )
-
-    def _parametrize_prepared_statements(self, prepared_statements, batch):
-        prepared_statements_dict = {}
-        serving_keys = set()
-        feature_name_order_by_psp = dict()
-        prefix_by_serving_index = {}
-        for prepared_statement in prepared_statements:
-            if prepared_statement.feature_group_id in self._skip_fg_ids:
-                continue
-            query_online = str(prepared_statement.query_online).replace("\n", " ")
-            prefix_by_serving_index[prepared_statement.prepared_statement_index] = (
-                prepared_statement.prefix
-            )
-
-            # In java prepared statement `?` is used for parametrization.
-            # In sqlalchemy `:feature_name` is used instead of `?`
-            pk_names = [
-                psp.name
-                for psp in sorted(
-                    prepared_statement.prepared_statement_parameters,
-                    key=lambda psp: psp.index,
-                )
-            ]
-            feature_name_order_by_psp[prepared_statement.prepared_statement_index] = (
-                dict(
-                    [
-                        (psp.name, psp.index)
-                        for psp in prepared_statement.prepared_statement_parameters
-                    ]
-                )
-            )
-            # construct serving key if it is not provided.
-            if self._serving_keys is None:
-                for pk_name in pk_names:
-                    # should use `ignore_prefix=True` because before hsfs 3.3,
-                    # only primary key of a feature group are matched in
-                    # user provided entry.
-                    serving_key = ServingKey(
-                        feature_name=pk_name,
-                        join_index=prepared_statement.prepared_statement_index,
-                        prefix=prepared_statement.prefix,
-                        ignore_prefix=True,
-                    )
-                    serving_keys.add(serving_key)
-
-            if not batch:
-                for pk_name in pk_names:
-                    query_online = self._parametrize_query(pk_name, query_online)
-                query_online = sql.text(query_online)
-            else:
-                query_online = self._parametrize_query("batch_ids", query_online)
-                query_online = sql.text(query_online)
-                query_online = query_online.bindparams(
-                    batch_ids=bindparam("batch_ids", expanding=True)
-                )
-
-            prepared_statements_dict[prepared_statement.prepared_statement_index] = (
-                query_online
-            )
-
-        # assign serving key if it is not provided.
-        if self._serving_keys is None:
-            self._serving_keys = serving_keys
-
-        return (
-            prepared_statements_dict,
-            feature_name_order_by_psp,
-            prefix_by_serving_index,
-        )
-
-    def _validate_serving_key(self, entry):
-        for key in entry:
-            if key not in self._valid_serving_key:
-                raise ValueError(
-                    f"'{key}' is not a correct serving key. Expect one of the"
-                    f" followings: [{', '.join(self._valid_serving_key)}]"
-                )
 
     def get_feature_vector(
         self,
@@ -546,7 +378,7 @@ class VectorServer:
             return polars_df
         else:
             raise ValueError(
-                "Unknown return type. Supported return types are 'pandas' and 'dict'"
+                "Unknown return type. Supported return types are 'pandas', 'polars' and 'dict'"
             )
 
     def get_inference_helpers(
@@ -586,137 +418,8 @@ class VectorServer:
             return polars_df
         else:
             raise ValueError(
-                "Unknown return type. Supported return types are 'dict' and 'pandas'"
+                "Unknown return type. Supported return types are 'pandas', 'polars' and 'dict'"
             )
-
-    def _vector_result(self, entry, prepared_statement_objects):
-        """Retrieve single vector with parallel queries using aiomysql engine."""
-
-        if all([isinstance(val, list) for val in entry.values()]):
-            raise ValueError(
-                "Entry is expected to be single value per primary key. "
-                "If you have already initialised prepared statements for single vector and now want to retrieve "
-                "batch vector please reinitialise prepared statements with  "
-                "`training_dataset.init_prepared_statement()` "
-                "or `feature_view.init_serving()`"
-            )
-        self._validate_serving_key(entry)
-        # Initialize the set of values
-        serving_vector = {}
-        bind_entries = {}
-        prepared_statement_execution = {}
-        for prepared_statement_index in prepared_statement_objects:
-            pk_entry = {}
-            next_statement = False
-            for sk in self._serving_key_by_serving_index[prepared_statement_index]:
-                if sk.required_serving_key not in entry.keys():
-                    # Check if there is any entry matched with feature name.
-                    if sk.feature_name in entry.keys():
-                        pk_entry[sk.feature_name] = entry[sk.feature_name]
-                    else:
-                        # User did not provide the necessary serving keys, we expect they have
-                        # provided the necessary features as passed_features.
-                        # We are going to check later if this is true
-                        next_statement = True
-                        break
-                else:
-                    pk_entry[sk.feature_name] = entry[sk.required_serving_key]
-            if next_statement:
-                continue
-            bind_entries[prepared_statement_index] = pk_entry
-            prepared_statement_execution[prepared_statement_index] = (
-                prepared_statement_objects[prepared_statement_index]
-            )
-
-        # run all the prepared statements in parallel using aiomysql engine
-        loop = asyncio.get_event_loop()
-        results_dict = loop.run_until_complete(
-            self._execute_prep_statements(prepared_statement_execution, bind_entries)
-        )
-
-        for key in results_dict:
-            for row in results_dict[key]:
-                result_dict = self.deserialize_complex_features(dict(row))
-                serving_vector.update(result_dict)
-
-        return serving_vector
-
-    def _batch_vector_results(self, entries, prepared_statement_objects):
-        """Execute prepared statements in parallel using aiomysql engine."""
-
-        # create dict object that will have of order of the vector as key and values as
-        # vector itself to stitch them correctly if there are multiple feature groups involved. At this point we
-        # expect that backend will return correctly ordered vectors.
-        batch_results = [{} for _ in range(len(entries))]
-        entry_values = {}
-        serving_keys_all_fg = []
-        prepared_stmts_to_execute = {}
-        # construct the list of entry values for binding to query
-        for prepared_statement_index in prepared_statement_objects:
-            # prepared_statement_index include fg with label only
-            # But _serving_key_by_serving_index include the index when the join_index is 0 (left side)
-            if prepared_statement_index not in self._serving_key_by_serving_index:
-                continue
-
-            prepared_stmts_to_execute[prepared_statement_index] = (
-                prepared_statement_objects[prepared_statement_index]
-            )
-            entry_values_tuples = list(
-                map(
-                    lambda e: tuple(
-                        [
-                            (
-                                e.get(sk.required_serving_key)
-                                # Check if there is any entry matched with feature name,
-                                # if the required serving key is not provided.
-                                or e.get(sk.feature_name)
-                            )
-                            for sk in self._serving_key_by_serving_index[
-                                prepared_statement_index
-                            ]
-                        ]
-                    ),
-                    entries,
-                )
-            )
-            entry_values[prepared_statement_index] = {"batch_ids": entry_values_tuples}
-
-        # run all the prepared statements in parallel using aiomysql engine
-        loop = asyncio.get_event_loop()
-        parallel_results = loop.run_until_complete(
-            self._execute_prep_statements(prepared_stmts_to_execute, entry_values)
-        )
-
-        # construct the results
-        for prepared_statement_index in prepared_stmts_to_execute:
-            statement_results = {}
-            serving_keys = self._serving_key_by_serving_index[prepared_statement_index]
-            serving_keys_all_fg += serving_keys
-            # Use prefix from prepare statement because prefix from serving key is collision adjusted.
-            prefix_features = [
-                (self._prefix_by_serving_index[prepared_statement_index] or "")
-                + sk.feature_name
-                for sk in self._serving_key_by_serving_index[prepared_statement_index]
-            ]
-            # iterate over results by index of the prepared statement
-            for row in parallel_results[prepared_statement_index]:
-                row_dict = dict(row)
-                # can primary key be complex feature?
-                result_dict = self.deserialize_complex_features(row_dict)
-                # note: should used serialized value
-                # as it is from users' input
-                statement_results[self._get_result_key(prefix_features, row_dict)] = (
-                    result_dict
-                )
-
-            # add partial results to the global results
-            for i, entry in enumerate(entries):
-                batch_results[i].update(
-                    statement_results.get(
-                        self._get_result_key_serving_key(serving_keys, entry), {}
-                    )
-                )
-        return batch_results, serving_keys_all_fg
 
     def get_complex_feature_schemas(self):
         return {
@@ -738,24 +441,6 @@ class VectorServer:
                 decoder = avro.io.BinaryDecoder(bytes_reader)
                 row_dict[feature_name] = schema.read(decoder)
         return row_dict
-
-    def refresh_mysql_connection(self):
-        try:
-            with self._prepared_statement_engine.connect():
-                pass
-        except exc.OperationalError:
-            self._set_mysql_connection()
-
-    def _make_preview_statement(self, statement, n):
-        return text(statement.text[: statement.text.find(" WHERE ")] + f" LIMIT {n}")
-
-    def _set_mysql_connection(self, options=None):
-        online_conn = self._storage_connector_api.get_online_connector(
-            self._feature_store_id
-        )
-        self._prepared_statement_engine = util.create_mysql_engine(
-            online_conn, self._external, options=options
-        )
 
     def _generate_vector(self, result_dict, fill_na=False):
         # feature values
@@ -786,40 +471,6 @@ class VectorServer:
                 ].transformation_fn
                 row_dict[feature_name] = transformation_fn(row_dict[feature_name])
         return row_dict
-
-    @staticmethod
-    def _get_result_key(primary_keys, result_dict):
-        result_key = []
-        for pk in primary_keys:
-            result_key.append(result_dict.get(pk))
-        return tuple(result_key)
-
-    @staticmethod
-    def _get_result_key_serving_key(serving_keys, result_dict):
-        result_key = []
-        for sk in serving_keys:
-            result_key.append(
-                result_dict.get(sk.required_serving_key)
-                or result_dict.get(sk.feature_name)
-            )
-        return tuple(result_key)
-
-    @staticmethod
-    def _parametrize_query(name, query_online):
-        # Now we have ordered pk_names, iterate over it and replace `?` with `:feature_name` one by one.
-        # Regex `"^(.*?)\?"` will identify 1st occurrence of `?` in the sql string and replace it with provided
-        # feature name, e.g. `:feature_name`:. As we iteratively update `query_online` we are always aiming to
-        # replace 1st occurrence of `?`. This approach can only work if primary key names are sorted properly.
-        # Regex `"^(.*?)\?"`:
-        # `^` - asserts position at start of a line
-        # `.*?` - matches any character (except for line terminators). `*?` Quantifier —
-        # Matches between zero and unlimited times, expanding until needed, i.e 1st occurrence of `\?`
-        # character.
-        return re.sub(
-            r"^(.*?)\?",
-            r"\1:" + name,
-            query_online,
-        )
 
     def _get_transformation_fns(self, entity):
         # get attached transformation functions
@@ -874,86 +525,6 @@ class VectorServer:
         )
         return transformation_fns
 
-    def filter_entry_by_join_index(self, entry, join_index):
-        fg_entry = {}
-        complete = True
-        for sk in self._serving_key_by_serving_index[join_index]:
-            fg_entry[sk.feature_name] = entry.get(sk.required_serving_key) or entry.get(
-                sk.feature_name
-            )  # fallback to use raw feature name
-            if fg_entry[sk.feature_name] is None:
-                complete = False
-                break
-        return complete, fg_entry
-
-    async def _set_aiomysql_connection(self, default_min_size: int, options=None):
-        online_connector = self._storage_connector_api.get_online_connector(
-            self._feature_store_id
-        )
-        self._async_pool = await util.create_async_engine(
-            online_connector, self._external, default_min_size, options=options
-        )
-
-    async def _query_async_sql(self, stmt, bind_params):
-        """Query prepared statement together with bind params using aiomysql connection pool"""
-        # Get a connection from the pool
-        async with self._async_pool.acquire() as conn:
-            # Execute the prepared statement
-            cursor = await conn.execute(stmt, bind_params)
-            # Fetch the result
-            resultset = await cursor.fetchall()
-            await cursor.close()
-
-        return resultset
-
-    async def _execute_prep_statements(self, prepared_statements: dict, entries: dict):
-        """Iterate over prepared statements to create async tasks
-        and gather all tasks results for a given list of entries."""
-
-        # validate if prepared_statements and entries have the same keys
-        if prepared_statements.keys() != entries.keys():
-            # iterate over prepared_statements and entries to find the missing key
-            # remove missing keys from prepared_statements
-            for key in list(prepared_statements.keys()):
-                if key not in entries:
-                    prepared_statements.pop(key)
-
-        tasks = [
-            asyncio.ensure_future(
-                self._query_async_sql(prepared_statements[key], entries[key])
-            )
-            for key in prepared_statements
-        ]
-        # Run the queries in parallel using asyncio.gather
-        results = await asyncio.gather(*tasks)
-        results_dict = {}
-        # Create a dict of results with the prepared statement index as key
-        for i, key in enumerate(prepared_statements):
-            results_dict[key] = results[i]
-        # TODO: close connection? closing connection pool here will make un-acquirable for future connections
-        # unless init_serving is explicitly called before get_feature_vector(s).
-        return results_dict
-
-    @property
-    def prepared_statement_engine(self):
-        """JDBC connection engine to retrieve connections to online features store from."""
-        return self._prepared_statement_engine
-
-    @prepared_statement_engine.setter
-    def prepared_statement_engine(self, prepared_statement_engine):
-        self._prepared_statement_engine = prepared_statement_engine
-
-    @property
-    def prepared_statements(self):
-        """The dict object of prepared_statements as values and kes as indices of positions in the query for
-        selecting features from feature groups of the training dataset.
-        """
-        return self._prepared_statements
-
-    @prepared_statements.setter
-    def prepared_statements(self, prepared_statements):
-        self._prepared_statements = prepared_statements
-
     @property
     def required_serving_keys(self):
         """Set of primary key names that is used as keys in input dict object for `get_feature_vector` method."""
@@ -1002,7 +573,7 @@ class VectorServer:
             and self._init_online_store_rest_client is False
         ):
             raise ValueError(
-                f"Default Online Store cCient is set to {self.DEFAULT_ONLINE_STORE_REST_CLIENT} but Online Store REST client"
+                f"Default Online Store Cient is set to {self.DEFAULT_ONLINE_STORE_REST_CLIENT} but Online Store REST client"
                 + " is not initialised. Call `init_serving` with init_client set to True before using it."
             )
         elif (
